% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_M4.R
\docType{data}
\name{model_M4}
\alias{model_M4}
\title{The pretrained model used in the M4 competition submission}
\format{A `xgb.Booster` object}
\usage{
model_M4
}
\description{
A xgboost model used for computing the weights of the combination of different forecasting methods
}
\keyword{datasets}
